Making a package for main
- Keeping all main.go's contents in 1 file was getting unwieldy
- Putting multiple source files in gohttp/ (package main) means `go
  build` commands need to take in multiple source files.
- Putting all of main's code into gohttp/main creates a second package
  `main`, and that just sounds confusing
- Putting all main-related packages in gohttp/main/cmd seems a good
  alternative to split up main's types into multiple files while still
  maintaining a system where all dependencies point away from main.
